requirements.txt
requests>=2.28
dnspython>=2.3
python-whois>=0.7
beautifulsoup4>=4.12
urllib3>=1.26
tqdm>=4.65
nmap>=0.7.1
shodan>=1.27.0
Jinja2>=3.1
wappalyzer-python>=0.4.0
